requirements.txt
anthropic>=0.40.0
beautifulsoup4>=4.12.0
matplotlib>=3.8.0
pandas>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
jupyter>=1.0.0
